Revert "Move utmp logic to Display"
This reverts commit 4c99b784a99db8028ca438edb1f4ddb03b3b6e8c.
