+++ src/auth/Auth.h
@@ -96,8 +96,6 @@ namespace SDDM {
         const QString &cookie() const;
         const QString &user() const;
         const QString &session() const;
-        qint64 sessionPid() const;
-
         AuthRequest *request();
         /**
          * True if an authentication or session is in progress
@@ -199,7 +197,7 @@ namespace SDDM {
         *
         * @param success true if succeeded
         */
-        void sessionStarted(bool success, qint64 pid);
+        void sessionStarted(bool success);
 
         /**
          * Emitted when the display server is ready.
