+++ src/helper/HelperApp.h
@@ -47,8 +47,8 @@ namespace SDDM {
         void info(const QString &message, Auth::Info type);
         void error(const QString &message, Auth::Error type);
         QProcessEnvironment authenticated(const QString &user);
-        void sessionOpened(bool success, qint64 pid);
         void displayServerStarted(const QString &displayName);
+        void sessionOpened(bool success);
 
     private slots:
         void setUp();
@@ -65,6 +65,23 @@ namespace SDDM {
         // TODO: get rid of this in a nice clean way along the way with moving to user session X server
         QString m_cookie { };
 
+        /*!
+         \brief Write utmp/wtmp/btmp records when a user logs in
+         \param vt  Virtual terminal (tty7, tty8,...)
+         \param displayName  Display (:0, :1,...)
+         \param user  User logging in
+         \param pid  User process ID (e.g. PID of startkde)
+         \param authSuccessful  Was authentication successful
+        */
+        void utmpLogin(const QString &vt, const QString &displayName, const QString &user, qint64 pid, bool authSuccessful);
+
+        /*!
+         \brief Write utmp/wtmp records when a user logs out
+         \param vt  Virtual terminal (tty7, tty8,...)
+         \param displayName  Display (:0, :1,...)
+         \param pid  User process ID (e.g. PID of startkde)
+        */
+        void utmpLogout(const QString &vt, const QString &displayName, qint64 pid);
     };
 }
